+++ Sources/StreamChatSwiftUI/ChatChannelList/ChatChannelListView.swift
@@ -234,6 +234,7 @@ public struct ChatChannelListContentView<Factory: ViewFactory>: View {
                     channels: viewModel.channels,
                     selectedChannel: $viewModel.selectedChannel,
                     swipedChannelId: $viewModel.swipedChannelId,
+                    scrolledChannelId: $viewModel.scrolledChannelId,
                     onlineIndicatorShown: viewModel.onlineIndicatorShown(for:),
                     imageLoader: channelHeaderLoader.image(for:),
                     onItemTap: onItemTap,