+++ Sources/StreamChatSwiftUI/ChatChannel/Gallery/GalleryView.swift
@@ -168,6 +168,7 @@ struct StreamVideoPlayer: View {
             }
         }
         .onAppear {
+            guard avPlayer == nil else { return }
             fileCDN.adjustedURL(for: url) { result in
                 switch result {
                 case let .success(url):
@@ -178,6 +179,9 @@ struct StreamVideoPlayer: View {
                     self.error = error
                 }
             }
+        }
+        .onDisappear {
+            avPlayer?.pause()
         }
     }
 }